requirements.txt
aws-cdk.core==1.129.0
aws-cdk.aws_iam==1.129.0
aws-cdk.aws_sqs==1.129.0
aws-cdk.aws_s3==1.129.0
aws-cdk.aws_lambda==1.129.0
aws-cdk.aws_lambda_event_Sources==1.129.0
constructs>=3.0.4,<4.0.0
